Create a recipe/explore & make /explore global
There's more to do for sure with both of these pages - in particular,
/explore, which is now effectively hidden.
